# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API Documentation",
        "description": "Test task API Documentation Description",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/api/branches": {
            "get": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branches list",
                "description": "Branches list.",
                "operationId": "branchList",
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/branch/{id}": {
            "get": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branch show",
                "description": "Branch show.",
                "operationId": "BranchShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/branch/create": {
            "post": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branch Create",
                "description": "Branch add.",
                "operationId": "BranchCreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Branch",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Branch"
                    },
                    {
                        "name": "brand",
                        "in": "query",
                        "description": "Brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Brand"
                    },
                    {
                        "name": "region",
                        "in": "query",
                        "description": "Region",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Region"
                    },
                    {
                        "name": "district",
                        "in": "query",
                        "description": "District",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "District"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/branch/update/{id}": {
            "put": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branch update",
                "description": "Branch update.",
                "operationId": "BranchUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Branch",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Branch"
                    },
                    {
                        "name": "brand",
                        "in": "query",
                        "description": "Brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Brand"
                    },
                    {
                        "name": "region",
                        "in": "query",
                        "description": "Region",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Region"
                    },
                    {
                        "name": "district",
                        "in": "query",
                        "description": "District",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "District"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/branch/delete/{id}": {
            "delete": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branch delete",
                "description": "Branch delete info.",
                "operationId": "BranchDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Delete error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/countOfBranch/{brand}": {
            "get": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "find Branch Location Count",
                "description": "Branch info location and count.",
                "operationId": "findBranchLocationCount",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "path",
                        "description": "Brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Brand"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}